+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Playlist Tools: added default value for duplicates selection and smart shuffle bias.
 - Music Map: added default value for duplicates selection and smart shuffle bias.
 - Display TF button: now supports editable presets. 8 built-in presets are now available by default.
+- Display TF button: added configurable actions on button click. Some of the actions are compatible with [Library-Tree-v2.4.0.mod.11](TODO) or later.
 - Presets: added Rating and Loved/Hated UI columns presets for DUI and CUI (at 'presets\UI').
 - Presets: new predefined presets for forced query filters.
 - Added support for [Biography-v1.4.2.mod.11](https://hydrogenaudio.org/index.php/topic,112914.msg1070993.html#msg1070993) similar artists database.